dd index fix
